Update Fedora releases for version compare CI
Since there is only one task in the strategy matrix, get rid of the
strategy matrix entirely, and move the sequence of calls to the last run
entry.

Split and comment the actions for the -sys and the main part.
Do the '-sys' part first, because the 'main' part depends on it.

For each, run for 'rawhide', 'f37', and 'f36'.

Signed-off-by: mulhern <amulhern@redhat.com>