libblkid-rs-sys: Allow default features for bindgen dependency
These were previously removed because bindgen's default also included
its CLI, which in turn included all the clap dependencies.

With 0.71 this is no longer the case, and allowing the default features
merely add a dependency on log and on prettyplease, so presumably better
logging and nicer formatting of the generated code.

|   |   |-- log v0.4.26
|   |   |-- prettyplease v0.2.32
|   |   |   |-- proc-macro2 v1.0.94
|   |   |   |   `-- unicode-ident v1.0.18
|   |   |   `-- syn v2.0.100
|   |   |       |-- proc-macro2 v1.0.94 (*)
|   |   |       |-- quote v1.0.39
|   |   |       |   `-- proc-macro2 v1.0.94 (*)
|   |   |       `-- unicode-ident v1.0.18

Signed-off-by: mulhern <amulhern@redhat.com>